Ambiente_Python/requirements.txt
flask

# ==== Core/Utilidades ====
python-dotenv>=1.0
loguru>=0.7
rich>=13.7
tqdm>=4.66
typer>=0.12

# ==== Data stack (análise) ====
numpy>=1.26
pandas>=2.2
pyarrow>=16.1
fastparquet>=2024.5.0
openpyxl>=3.1
xlsxwriter>=3.2
scipy>=1.13
scikit-learn>=1.5
statsmodels>=0.14

# ==== Visualização ====
matplotlib>=3.8
plotly>=5.22
altair>=5.3

# ==== Bancos de dados / ORM ====
SQLAlchemy>=2.0
pymysql>=1.1            # MySQL/MariaDB (usado no seu OKR)
psycopg2-binary>=2.9    # PostgreSQL
pyodbc>=5.1             # SQL Server / ODBC

# ==== Arquivos, planilhas, documentos ====
python-docx>=1.1        # gerar DOCX
pdfplumber>=0.11        # ler PDFs
reportlab>=4.2          # gerar PDFs
fpdf2>=2.7              # gerar PDFs leve
pyyaml>=6.0

# ==== Web/HTTP, APIs, validação ====
requests>=2.32
httpx>=0.27
aiohttp>=3.9
tenacity>=8.3           # retries
pydantic>=2.7

# ==== Web scraping / parsing HTML ====
beautifulsoup4>=4.12
lxml>=5.2
readability-lxml>=0.8
html5lib>=1.1

# ==== Automação, Jobs e SSH ====
APScheduler>=3.10
schedule>=1.2
paramiko>=3.4
fabric>=3.2

# ==== Cloud & Big Data (opcional – comente se não usar) ====
google-cloud-storage>=2.18
google-cloud-bigquery>=3.25
pandas-gbq>=0.22
boto3>=1.34             # AWS

# ==== Serialização, performance, cache ====
orjson>=3.10
ujson>=5.10
cachetools>=5.4

# ==== Criptografia / segurança ====
cryptography>=43.0

# ==== E-mail (opcional) ====
yagmail>=0.15

# ==== Testes (opcional) ====
pytest>=8.3
